update interpreter implementation using an AST
