+++ pyproject.toml
@@ -46,6 +46,7 @@ skip-string-normalization = true
 [tool.ruff]
 target-version = "py310"
 line-length = 90
+ignore = ["E999"]  # SyntaxError from unsupported pattern matching
 #select = ["E", "F", "B", "Q"]
 #[tool.ruff.per-file-ignores]
 #"tests/*.py" = ["D205", "D400"]  # docstring formatting
