+++ tests/unit/test_interpreter.py
@@ -1,7 +1,7 @@
 from hypothesis import given
 from hypothesis import strategies as st
 
-from spi.calculator import interpret
+from spi.interpreter import interpret
 
 
 @given(st.integers(0))
@@ -28,11 +28,17 @@ def test_interpret_integer_pair(x: int, y: int, op: str) -> None:
 @given(st.lists(st.integers(1), min_size=5, max_size=5))
 def test_interpret_full_expression(xs: list[int]) -> None:
     text = f'{xs[0]} + {xs[1]} * {xs[2]} - {xs[3]} / {xs[4]}'
-    assert interpret(text) == xs[0] + xs[1] * xs[2] - xs[3] // xs[4]
+    try:
+        assert interpret(text) == xs[0] + xs[1] * xs[2] - xs[3] // xs[4]
+    except ZeroDivisionError:
+        pass
 
 
 @given(st.lists(st.integers(1), min_size=7, max_size=7))
 def test_interpret_parenthesis_expression(xs: list[int]) -> None:
     text = f'{xs[0]} + {xs[1]} * ({xs[2]} / ({xs[3]} / ({xs[4]} + {xs[5]}) - {xs[6]}))'
     expected = xs[0] + xs[1] * (xs[2] // (xs[3] // (xs[4] + xs[5]) - xs[6]))
-    assert interpret(text) == expected
+    try:
+        assert interpret(text) == expected
+    except ZeroDivisionError:
+        pass
